requirements.txt
asgiref==3.5.2
boto3==1.25.1
botocore==1.28.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
cryptography==37.0.3
defusedxml==0.7.1
distlib==0.3.6
Django==4.0.5
django-cors-headers==3.13.0
django-debug-toolbar==3.4.0
django-filter==21.1
django-storages==1.13.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
# djangorestframework-simplejwt==5.2.2
djoser==2.1.0
drf-nested-routers==0.93.4
filelock==3.8.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.1.1
pipenv==2022.10.12
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.16.5
virtualenv-clone==0.5.7
whitenoise==6.2.0
drf-yasg==1.21.7

